more refinements
