--- spring-cloud-kubernetes-istio/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,16 +1,0 @@
-{
-	"properties": [
-		{
-			"name": "spring.cloud.istio.client.envoyPort",
-			"type": "java.lang.Integer",
-			"description": "Envoy Proxy Port to check if Istio is installed",
-			"defaultValue": 15090
-		},
-		{
-			"name": "spring.cloud.istio.client.testPath",
-			"type": "java.lang.String",
-			"description": "Test URL path to check if Istio is enabled where the pod is running",
-			"defaultValue": "stats/prometheus"
-		}
-	]
-}
